renamed overlayImage, updated example data
The previous overlay was actually used as a backdrop to the similarity
map and therefore not overlaid at all.

Also added a new overlayImage which will actually be on top of the
similarity map. For example, this could be an edge image of a bright
field scan.

The number of such additional images is in principle not limited. The
final input format and parsing should work with a list of those images
and their parameters.